requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# MongoDB
pymongo>=4.0.0

# HTTP client
httpx>=0.24.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Cross-encoder reranking
sentence-transformers>=2.2.0

# LLM Providers
google-generativeai>=0.3.0
openai>=1.0.0

# Utilities
python-dateutil>=2.8.0
requests>=2.28.0
